✨ feat: 환불 관련 배송비 로직 구현

+++ src/main/java/com/jajaja/domain/point/repository/PointRepository.java
@@ -6,7 +6,7 @@ import com.jajaja.domain.point.entity.enums.PointType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface PointRepository extends JpaRepository<Point, Integer>, PointRepositoryCustom {
+public interface PointRepository extends JpaRepository<Point, Long>, PointRepositoryCustom {
 
     boolean existsByTypeAndOrderProduct(PointType type, OrderProduct orderProduct);
